chore(Config) Move configuration utils into configuration
