+++ nes-single-node-worker/include/Configuration.hpp
@@ -44,46 +44,4 @@ public:
     Configurations::WorkerConfiguration engineConfiguration = {"engineConfiguration", "NodeEngine Configuration"};
     Configurations::QueryCompilerConfiguration queryCompilerConfiguration = {"queryCompilerConfiguration", "QueryCompiler Configuration"};
 };
-
-template <typename T>
-void generateHelp(std::ostream& ostream)
-{
-    T config;
-    Configurations::PrintingVisitor visitor{ostream};
-    config.accept(visitor);
-}
-
-///TODO(#130): Generalize and move into `nes-configuration`
-/// CLI > ConfigFile
-template <typename T>
-auto loadConfiguration(const int argc, const char** argv)
-{
-    /// Convert the POSIX command line arguments to a map of strings.
-    std::map<std::string, std::string> commandLineParams;
-    for (int i = 1; i < argc; ++i)
-    {
-        const size_t pos = std::string(argv[i]).find('=');
-        const std::string arg{argv[i]};
-        if (arg == "--help")
-        {
-            generateHelp<T>(std::cout);
-            std::exit(0);
-        }
-        commandLineParams.insert({arg.substr(0, pos), arg.substr(pos + 1, arg.length() - 1)});
-    }
-
-    /// Create a configuration object with default values.
-    T config;
-
-    /// Read options from the YAML file.
-    if (const auto configPath = commandLineParams.find("--" + Configurations::CONFIG_PATH); configPath != commandLineParams.end())
-    {
-        config.overwriteConfigWithYAMLFileInput(configPath->second);
-    }
-
-    /// Options specified on the command line have the highest precedence.
-    config.overwriteConfigWithCommandLineInput(commandLineParams);
-
-    return config;
-}
 }
